Conflicts:
	cloud/bin/clean_instance_for_imaging.sh

+++ lib/WormBase/Update/Production/DeployWebapp.pm
@@ -18,6 +18,7 @@ has 'step' => (
 );
 
 
+
 has 'pm_version' => (
     is => 'ro',
     lazy_build => 1
@@ -32,18 +33,6 @@ sub _build_pm_version {
     return $software_version;
 }
 
-has 'hg_revision' => (
-    is => 'ro',
-    lazy_build => 1,
-    );
-
-sub _build_hg_revision {       
-    my $self = shift;
-    chdir($self->app_staging_dir);
-    my $hg_revision = `hg tip --template '{rev}'`;
-    chomp $hg_revision;
-    return $hg_revision;
-}
 
 has 'git_commits' => (
     is => 'ro',
@@ -80,7 +69,12 @@ has 'app_version' => (
 
 sub _build_app_version {
     my $self = shift;
+
+    # Getting the version from localhost doesn't work. It may be running a version > than production.
+    # Maybe get the current production version from rest instead?
+    # In THIS case, we are using the supplied release.
     my $release = $self->release;  
+
     my $software_version  = $self->pm_version;
 #    my $software_revision = $self->git_cumulative_commits;
     my $software_revision = $self->git_commits;
@@ -94,9 +88,7 @@ sub _build_app_version {
 ########################################################
 
 sub run {
-    my $self = shift;           
-    my $release = $self->release;
-
+    my $self = shift;
     $self->dump_version_file;
     $self->rsync_staging_directory;    
 #    $self->create_environment_file;
@@ -112,7 +104,11 @@ sub dump_version_file {
     $self->log->info('dumping software version file');
 
     chdir($self->app_staging_dir);
-    my $release = $self->release;  
+
+    # Getting the version from localhost doesn't work. It may be running a version > than production.
+    # Maybe get the current production version from rest instead?
+#    my $release = $self->production_release;  
+
     my $software_version = $self->pm_version;
     my $software_commits = $self->git_commits;
     my $software_commits_running = $self->git_cumulative_commits;
@@ -125,15 +121,14 @@ sub dump_version_file {
     open OUT,">VERSION.txt";
     print OUT <<END;
 DATE=$date
-DATABASE_VERSION=$release
 SOFTWARE_VERSION=$software_version
 COMMITS_SINCE_TAG=$software_commits
 COMMITS_CUMULATIVE=$software_commits_running
 VERSION_STRING="v{$software_version}r${software_commits}"
 END
 close OUT;
-
 }
+#DATABASE_VERSION=$release
 
 
 # The WormBase environment file
@@ -225,19 +220,25 @@ sub rsync_staging_directory {
 
 #    foreach my $node (@$remote_nodes) {
     foreach my $node (@$local_nodes,@$remote_nodes) {
-	$self->log->debug("rsync staging to $node");
+	$self->log->info("   deploying the staging directory to $node");
 	my $ssh = $self->ssh($node);
 	$ssh->error && $self->log->logdie("Can't ssh to $node: " . $ssh->error);
 
-	$ssh->system("mkdir $app_root/website/$app_version") or $self->log->logdie("Couldn't create a new app version on $node: " . $ssh->error);
-
-	$self->system_call("rsync -Ca --exclude logs --exclude tmp --exclude .hg --exclude extlib.tgz --exclude wormbase.env --exclude extlib $staging_dir/ ${node}:$app_root/website/$app_version",'rsyncing staging directory into production');
-
-
-	# Update the symlink.  Here or part of GoLive?
-	$ssh->system("cd $app_root/website; mkdir $app_root/website/$app_version/logs ; chmod 777 $app_root/website/$app_version/logs ; rm production;  ln -s $app_version production")
-	    or $self->log->logdie("Couldn't update the production symlink");
-		
+	# Approach 1: each deployment push is a new directory.
+	if (0) {
+	    $ssh->system("mkdir $app_root/website/$app_version") or $self->log->warn("Couldn't create a new app version on $node: " . $ssh->error);
+	    $self->system_call("rsync -Ca --exclude logs --exclude tmp --exclude .hg --exclude extlib.tgz --exclude wormbase.env --exclude extlib $staging_dir/ ${node}:$app_root/website/$app_version",'rsyncing staging directory into production');
+	    # Update the symlink.  Here or part of GoLive?
+	    $ssh->system("cd $app_root/website; mkdir $app_root/website/$app_version/logs ; chmod 777 $app_root/website/$app_version/logs ; rm production;  ln -s $app_version production")
+		or $self->log->warn("Couldn't update the production symlink");	    
+	}
+	
+	# Approach 2: simply rsync to the production directory.
+	# First back up production.
+	$ssh->system("cp -r /usr/local/wormbase/website/production /usr/local/wormbase/website/archive/$app_version")
+	    or $self->log->warn("Couldn't back up the current production directory to $node:website/archive/$app_version");
+	$self->system_call("rsync -Ca --exclude logs --exclude tmp --exclude .hg --exclude extlib.tgz --exclude wormbase.env --exclude extlib $staging_dir/ ${node}:$app_root/website/production",'rsyncing staging directory into production');		
+	$self->send_hup_to_starman($ssh,$node);
     }
 }
 
@@ -262,9 +263,19 @@ sub save_production_reference {
     $self->log->info('saving production reference');
     my $wormbase_root = $self->wormbase_root;
     chdir("$wormbase_root/website");
-    $self->system_call("rm -rf production",'removing the old production version');
-    $self->system_call("cp -r staging production",'saving the new production version');
+    my $staging_dir = $self->app_staging_dir;
+    $self->system_call("rm -rf production.current",'removing the old production version');
+    $self->system_call("cp -r $staging_dir production.current",'saving a reference of the current production version');
 }
+
+
+sub send_hup_to_starman {
+    my ($self,$ssh,$node) = @_;
+    $self->log->info("   sending HUP to starman on $node");
+    $ssh->system("kill -HUP `head -1 /tmp/production.pid`");
+}
+
+
 
 
 1;
